requirements.txt
pandas
beautifulsoup4
selenium
python-dotenv
tqdm
